% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R, R/generics.R
\name{as.booltype.default}
\alias{as.booltype.default}
\alias{as.booltype}
\title{Coerce to booltype (generic)}
\usage{
\method{as.booltype}{default}(x, booltype = "logical", ...)

as.booltype(x, booltype, ...)
}
\arguments{
\item{x}{object to coerce}

\item{booltype}{target \code{\link[=booltype]{booltype()}} given as integer or as character}

\item{...}{further arguments}
}
\value{
\code{x} coerced to \code{booltype}
}
\description{
Coerce to booltype (generic)
}
\section{Methods (by class)}{
\itemize{
\item \code{as.booltype(default)}: default method for as.booltype

}}
\examples{
as.booltype(0:1)
as.booltype(0:1, "logical")
as.booltype(0:1, "bit")
as.booltype(0:1, "bitwhich")
as.booltype(0:1, "which", maxindex=2)
as.booltype(0:1, "ri")
}
\seealso{
\code{\link{CoercionToStandard}}, \code{\link[=booltypes]{booltypes()}}, \code{\link[=booltype]{booltype()}},
\code{\link[=is.booltype]{is.booltype()}}
}
