% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R
\name{as.character.bit}
\alias{as.character.bit}
\title{Coerce bit to character}
\usage{
\method{as.character}{bit}(x, ...)
}
\arguments{
\item{x}{a \code{\link{bit}} vector}

\item{...}{ignored}
}
\value{
a character vector of zeroes and ones
}
\description{
Coerce bit to character
}
\examples{
as.character(bit(12))
}
