% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R
\name{as.character.bitwhich}
\alias{as.character.bitwhich}
\title{Coerce bitwhich to character}
\usage{
\method{as.character}{bitwhich}(x, ...)
}
\arguments{
\item{x}{a \code{\link[=bitwhich]{bitwhich()}} vector}

\item{...}{ignored}
}
\value{
a character vector of zeroes and ones
}
\description{
Coerce bitwhich to character
}
\examples{
as.character(bitwhich(12))
}
