% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R, R/generics.R
\name{as.ri.ri}
\alias{as.ri.ri}
\alias{as.ri.default}
\alias{as.ri}
\title{Coerce to ri}
\usage{
\method{as.ri}{ri}(x, ...)

\method{as.ri}{default}(x, ...)

as.ri(x, ...)
}
\arguments{
\item{x}{object to coerce}

\item{...}{further arguments}
}
\value{
an \code{\link{ri}} object
}
\description{
Coerce to ri
}
\section{Methods (by class)}{
\itemize{
\item \code{as.ri(ri)}: method to coerce \code{\link{ri}} to \code{\link{ri}}

\item \code{as.ri(default)}: default method to coerce to \code{\link{ri}}

}}
\examples{
as.ri(c(FALSE, TRUE, FALSE, TRUE))
}
\seealso{
\code{\link{CoercionToStandard}}, \code{\link{as.booltype}}, \code{\link{as.bit}}, \code{\link{as.bitwhich}}
, \code{\link{as.which}}, \code{\link{as.ri}}, \code{\link[ff]{as.hi}},  \code{\link[ff]{as.ff}}
}
\author{
Jens Oehlschlägel
}
\keyword{classes}
\keyword{logic}
