\name{BayesIMTR-package}
\alias{BayesIMTR-package}
\alias{BayesIMTR}
\docType{package}
\title{
  Bayesian Identification of Master Transcriptional Regulators
}
\description{
  Using a Bayesian hierarchical model together with more than 10000 TF ChIP-seq peaks, the package is capable of identifying the possible master transcriptional regulators of input open region peakset(ATAC-seq,DNase-seq,etc...)
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Zeyu Lu,
Maintainer: Zeyu Lu <zeyul@smu.edu>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
