% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{BayesIMTR}
\alias{BayesIMTR}
\title{Title}
\usage{
BayesIMTR(
  file,
  format = c("bed", "narrowPeak", "broadPeak", "bigNarrowPeak"),
  N = 1000,
  bin_width = 1000
)
}
\arguments{
\item{file}{file path to the user-input.}

\item{format}{format can be .bed, .narrowPeak, .broadPeak and .bigNarrowPeak.}

\item{N}{number of rounds for gibbs sampler, recommended for > 1000.}

\item{bin_width}{desired width of bin, should be in 100/500/1000.}
}
\value{
A list object contains results of Gibbs sampler.
}
\description{
Title
}
\examples{
input_peak <- "./CTCF.bed"
gibbs_results <- BayesIMTR(input_peak, format = "bed", N = 1000, bin_width = 1000)
}
