% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{BIMTR}
\alias{BIMTR}
\title{BIMTR}
\usage{
BIMTR(
  file,
  format = c("bed", "narrowPeak", "broadPeak", "bigNarrowPeak", "csv"),
  N = 5000,
  bin_width = 1000
)
}
\arguments{
\item{file}{file path to the user-input.}

\item{format}{format can be .bed, .narrowPeak, .broadPeak and .bigNarrowPeak.}

\item{N}{number of iterations for gibbs sampler, recommended for >= 5000.}

\item{bin_width}{desired width of bin, should be in 100/500/1000.}
}
\value{
A list object contains results of Gibbs sampler.
}
\description{
Main interface to run BIMTR method, please set the input file path, input file format, number of iterations and bin width.
users can also change default parameters used in Gibbs sampler.
}
