% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{BIMTR_multi}
\alias{BIMTR_multi}
\title{BIMTR parallel computation.}
\usage{
BIMTR_multi(
  file,
  format = c("bed", "narrowPeak", "broadPeak", "bigNarrowPeak"),
  N = 2000,
  bin_width = 1000,
  numCores
)
}
\arguments{
\item{file}{file path to the user-input.}

\item{format}{format can be .bed, .narrowPeak, .broadPeak and .bigNarrowPeak.}

\item{N}{number of iterations for gibbs sampler, recommended for > 5000.}

\item{bin_width}{desired width of bin, should be in 100/500/1000.}

\item{numCores}{number of cores used in parallel computation.}
}
\value{
A list object of list objects that contains results of Gibbs sampler.
}
\description{
the multi-cores parallel version of BIMTR to acquire parallel MCMC chains for gelman-rubin diagnostic.
}
