% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R
\name{bit}
\alias{bit}
\title{Create empty bit vector}
\usage{
bit(length = 0L)
}
\arguments{
\item{length}{length in bits}
}
\value{
\code{bit} returns a vector of integer sufficiently long to store 'length' bits
}
\description{
Bit vectors are a boolean type wihout \code{NA} that requires by factor 32 less RAM than \code{\link{logical}}.
For details on usage see the \href{../doc/bit-usage.html}{usage-vignette} and for details on performance see \href{../doc/bit-performance.html}{performance-vignette}
}
\examples{
bit(12)
!bit(12)
str(bit(128))
}
\seealso{
\code{\link{booltype}}, \code{\link{bitwhich}}, \code{\link{logical}}
}
\keyword{classes}
\keyword{logic}
