% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{BIT_compare}
\alias{BIT_compare}
\title{BIT_compare}
\usage{
BIT_compare(
  file1,
  file2,
  output_path,
  show = TRUE,
  plot_scatter = TRUE,
  format = c(NULL, NULL),
  N = 5000,
  bin_width = 1000,
  burnin = NULL,
  genome = c("hg38", "mm10"),
  seed = 42
)
}
\arguments{
\item{file1}{file path to the user-input file 1.}

\item{file2}{file path to the user-input file 2.}

\item{output_path}{absoluate or relative directory to store the Gibbs sampler data.}

\item{show}{Whether to show the results table, default: TRUE}

\item{plot_scatter}{Whether to plot the BIT scores of two region sets in one scatter plot, default: TRUE}

\item{format}{if specify as NULL, BIT will automatically read and judge the file type based on extension, default: NULL.}

\item{N}{number of iterations for gibbs sampler, recommended for >= 5000, default: 5000.}

\item{bin_width}{desired width of bin, default: 1000.}

\item{genome}{Genome of the reference TR ChIP-seq data, must be "hg38" or "mm10"}

\item{seed}{Set random seed for reproducibility.}
}
\description{
compare BIT identifid TRs for two user input epigenomic region sets.
}
