% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTCF_Demo}
\alias{CTCF_Demo}
\title{CTCF Demo data}
\format{
## 'CTCF_Demo'
A data frame with 13511 rows and 4 columns
\describe{
  \item{TF}{transcription factor HGNC symbol.}
  \item{GOOD}{number of good cases of demo input ATAC-seq peakset.}
  \item{TOTAL}{number of total informative cases of demo input ATAC-seq peakset.}
  \item{Ratio}{ratio of good cases over all informative cases.}
}
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE153237
}
\usage{
CTCF_Demo
}
\description{
A transformed demo data set used for showing the workflow of BIMTR.
}
\details{
Xu B, Wang H, Wright S, Hyle J, Zhang Y, Shao Y, Niu M, Fan Y, Rosikiewicz W, Djekidel MN, Peng J, Lu R, Li C.
Acute depletion of CTCF rewires genome-wide chromatin accessibility. Genome Biol. 2021 Aug 24;22(1):244.
doi: 10.1186/s13059-021-02466-0. PMID: 34429148; PMCID: PMC8386078.
}
\keyword{datasets}
