% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{display_tables}
\alias{display_tables}
\title{To show the ranking table by inspecting the results of Gibbs sampler.}
\usage{
display_tables(dat, burnin = NULL)
}
\arguments{
\item{dat}{results from Main_sampling.}

\item{burnin}{number of samples used for burn-in. If not specify, BIT will use the half of the iterations as burn in.}
}
\value{
a data.frame object contains TR names, BIT scores and 95 CI for each TR.
}
\description{
To show the ranking table by inspecting the results of Gibbs sampler.
}
