% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{filter_peaks}
\alias{filter_peaks}
\title{filter peaks based on CREs, Promoter, Enhancer}
\usage{
filter_peaks(
  input_vec,
  option = c("ALL", "CREs", "CLS", "ELS"),
  bin_width = c(100, 500, 1000)
)
}
\arguments{
\item{input_vec}{input vector that contains indices from import_peaks}

\item{option}{option, ALL: Non-filtering, CREs: All cis-regulatory elements, PLS: Promoters, ELS: Enhancers.}

\item{bin_width}{bin width}
}
\value{
a vector contains filtered bin indices.
}
\description{
filter peaks based on CREs, Promoter, Enhancer
}
