% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{gelman_rubin}
\alias{gelman_rubin}
\title{gelman-rubin convergence diagnostic}
\usage{
gelman_rubin(
  file,
  format = c("bed", "narrowPeak", "broadPeak", "bigNarrowPeak"),
  N = 2000,
  bin_width = 1000,
  nchains = parallel::detectCores() - 1
)
}
\arguments{
\item{file}{file path to the user-input.}

\item{format}{format can be .bed, .narrowPeak, .broadPeak and .bigNarrowPeak.}

\item{N}{number of iterations for gibbs sampler, recommended for > 5000.}

\item{bin_width}{desired width of bin, should be in 100/500/1000.}

\item{nchains}{number of chains used in diagnostic.}
}
\value{
A list object contains multiple mcmc chains.
}
\description{
gelman-rubin convergence diagnostic based on parallel computation, used chains can be varied but must less or equal to the usable cores.
}
