% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demo.R
\name{gelman_rubin_Demo}
\alias{gelman_rubin_Demo}
\title{gelman-rubin demo}
\usage{
gelman_rubin_Demo(
  N = 2000,
  dat = c("CTCF", "ZBTB7A", "KDM1A"),
  nchains = parallel::detectCores() - 1
)
}
\arguments{
\item{N}{number of rounds}

\item{dat}{pre-processed Demo data, including CTCF, ZBTB7A and KDM1A. For details please check the manual for the data.}

\item{nchains}{number of chains.}
}
\value{
A list object contains gelman-rubin diagnostic results.
}
\description{
Gelman-Rubin convergence diagnostic for Demo dataset, recommend for N > 2000
}
