% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KDM1A_Demo}
\alias{KDM1A_Demo}
\title{KDM1A Demo data}
\format{
## 'KDM1A_Demo'
A data frame with 13511 rows and 4 columns
\describe{
  \item{TF}{transcription factor HGNC symbol.}
  \item{GOOD}{number of good cases of demo input ATAC-seq peakset.}
  \item{TOTAL}{number of total informative cases of demo input ATAC-seq peakset.}
  \item{Ratio}{ratio of good cases over all informative cases.}
}
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE140505
}
\usage{
KDM1A_Demo
}
\description{
A transformed demo data set used for testing the efficiency of BayesIMTR
and showing the workflow.
}
\details{
Park DE, Cheng J, McGrath JP, Lim MY, Cushman C, Swanson SK, Tillgren ML, Paulo JA, Gokhale PC, Florens L, Washburn MP, Trojer P, DeCaprio JA.
Author Correction: Merkel cell polyomavirus activates LSD1-mediated blockade of non-canonical BAF to regulate transformation and tumorigenesis.
Nat Cell Biol. 2020 Jun;22(6):752. doi: 10.1038/s41556-020-0533-9. Erratum for: Nat Cell Biol. 2020 May;22(5):603-615. PMID: 32415271; PMCID: PMC8107828.
}
\keyword{datasets}
