% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_chip_data}
\alias{load_chip_data}
\title{load the pre-compiled chip-seq data.}
\usage{
load_chip_data(data_path, bin_width, genome = c("hg38", "mm10"))
}
\arguments{
\item{data_path}{path to the ChIP-seq data folder, can be absolute or relative path.}

\item{bin_width}{width of bin, which should be in 100/500/1000 and map with your ChIP-seq data.}
}
\description{
load the pre-compiled chip-seq data. Please follow the tutorial on: https://github.com/ZeyuL01/BIT.
}
