% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{mcmc_list_transformer}
\alias{mcmc_list_transformer}
\title{mcmc_list_transformer}
\usage{
mcmc_list_transformer(dat, label, nchains)
}
\arguments{
\item{dat}{input results from parallel chains.}

\item{label}{label of TFs.}

\item{nchains}{number of chains.}
}
\value{
A list object contains multiple mcmc chains.
}
\description{
function to transform parallel chains into a mcmc list.
}
