% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Metadata}
\alias{Metadata}
\alias{is.sorted}
\alias{is.sorted<-}
\alias{na.count}
\alias{na.count<-}
\alias{nvalid}
\alias{nunique}
\alias{nunique<-}
\alias{nties}
\alias{nties<-}
\title{Generics related to cache access}
\usage{
is.sorted(x, ...)

is.sorted(x, ...) <- value

na.count(x, ...)

na.count(x, ...) <- value

nvalid(x, ...)

nunique(x, ...)

nunique(x, ...) <- value

nties(x, ...)

nties(x, ...) <- value
}
\arguments{
\item{x}{some object}

\item{...}{ignored}

\item{value}{value assigned on responsibility of the user}
}
\value{
see help of the available methods
}
\description{
These generics are packaged here for methods in packages \code{bit64} and
\code{ff}.
}
\details{
see help of the available methods
}
\examples{
  methods("na.count")
}
\seealso{
\code{\link[bit64:is.sorted.integer64]{bit64::is.sorted.integer64()}}, \code{\link[bit64:is.sorted.integer64]{bit64::na.count.integer64()}},
\code{\link[bit64:is.sorted.integer64]{bit64::nvalid.integer64()}}, \code{\link[bit64:is.sorted.integer64]{bit64::nunique.integer64()}}, \code{\link[bit64:is.sorted.integer64]{bit64::nties.integer64()}}
}
\author{
Jens Oehlschlägel \href{mailto:Jens.Oehlschlaegel@truecluster.com}{Jens.Oehlschlaegel@truecluster.com}
}
\keyword{environment}
\keyword{methods}
