% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PLS_list}
\alias{PLS_list}
\title{PLS_list}
\format{
## 'PLS_list'
A list with 3 vectors.
\describe{
  \item{PLS_1000}{candidate promoters transformed to indices with 1000 bin width}
  \item{PLS_500}{candidate promoters transformed to indices with 500 bin width}
  \item{PLS_100}{candidate promoters transformed to indices with 100 bin width}
}
}
\source{
https://www.nature.com/articles/s41586-020-2493-4
}
\usage{
PLS_list
}
\description{
A pre-processed list that contains indices of all candidate promoters with 3 type of bin width.
}
\details{
ENCODE Project Consortium, Jill E. Moore, Michael J. Purcaro, Henry E. Pratt, Charles B. Epstein, Noam Shoresh, Jessika Adrian, et al. 2020.
Expanded Encyclopaedias of DNA Elements in the Human and Mouse Genomes. Nature 583 (7818): 699–710.
doi: 10.1038/s41586-020-2493-4. Epub 2020 Jul 29. Erratum in: Nature. 2022 May;605(7909):E3. PMID: 32728249; PMCID: PMC7410828.
}
\keyword{datasets}
