% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R
\name{print.bit}
\alias{print.bit}
\title{Print method for bit}
\usage{
\method{print}{bit}(x, ...)
}
\arguments{
\item{x}{a bit vector}

\item{...}{passed to print}
}
\value{
a character vector showing first and last elements of the bit vector
}
\description{
Print method for bit
}
\examples{
print(bit(120))
}
