% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R
\name{print.bitwhich}
\alias{print.bitwhich}
\title{Print method for bitwhich}
\usage{
\method{print}{bitwhich}(x, ...)
}
\arguments{
\item{x}{a \code{\link[=bitwhich]{bitwhich()}} object}

\item{...}{ignored}
}
\description{
Print method for bitwhich
}
