% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R
\name{rev.booltype}
\alias{rev.booltype}
\alias{rev.bit}
\alias{rev.bitwhich}
\title{Reversing bit and bitwhich vectors}
\usage{
\method{rev}{bit}(x)

\method{rev}{bitwhich}(x)
}
\arguments{
\item{x}{bit or bitwhich object}
}
\value{
An object of class 'bit' or 'bitwhich'
}
\description{
Creating new bit or bitwhich by reversing such vectors
}
\examples{

 rev(as.bit(c(FALSE, TRUE)))
 rev(as.bitwhich(c(FALSE, TRUE)))
}
\seealso{
\code{\link[=rev]{rev()}}, \code{\link[=bit]{bit()}} , \code{\link[=bitwhich]{bitwhich()}}
}
\author{
Jens Oehlschlägel
}
\keyword{classes}
\keyword{logic}
