% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integerutil.R
\name{reverse}
\alias{reverse}
\title{Reverse atomic vector}
\usage{
reverse(x)
}
\arguments{
\item{x}{an R vector}
}
\value{
a reversed vector
}
\description{
Returns a reversed copy -- with attributes retained.
}
\details{
This is substantially faster than \code{\link{rev}}
}
\examples{
x <- factor(letters)
rev(x)
reverse(x)
\dontrun{
x <- 1:1e7
system.time(rev(x))
system.time(reverse(x))
}
}
\seealso{
\code{\link{rev}}, \code{\link{copy}}
}
