% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Show_Results}
\alias{Show_Results}
\title{To show the ranking table by inspecting the results of Gibbs sampler.}
\usage{
Show_Results(dat, burnin, centerized = FALSE)
}
\arguments{
\item{dat}{results from Main_sampling.}

\item{burnin}{number of samples used for burn-in.}
}
\value{
a list object contains two data.frame. one by ranking the theta_ij, the other one by ranking theta_i.
}
\description{
To show the ranking table by inspecting the results of Gibbs sampler.
}
