% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integerutil.R
\name{still.identical}
\alias{still.identical}
\title{Test for C-level identity of two atomic vectors}
\usage{
still.identical(x, y)
}
\arguments{
\item{x}{an atomic vector}

\item{y}{an atomic vector}
}
\value{
logical scalar
}
\description{
Test for C-level identity of two atomic vectors
}
\examples{
x <- 1:2
y <- x
z <- copy_vector(x)
still.identical(y,x)
still.identical(z,x)
}
