% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R
\name{str.bit}
\alias{str.bit}
\title{Str method for bit}
\usage{
\method{str}{bit}(
  object,
  vec.len = strO$vec.len,
  give.head = TRUE,
  give.length = give.head,
  ...
)
}
\arguments{
\item{object}{any \R object about which you want to have some
    information.}

\item{vec.len}{numeric (>= 0) indicating how many \sQuote{first few} elements
    are displayed of each vector.  The number is multiplied by different
    factors (from .5 to 3) depending on the kind of vector.  Defaults to
    the \code{vec.len} component of option \code{"str"} (see
    \code{\link{options}}) which defaults to 4.}

\item{give.head}{logical; if \code{TRUE} (default), give (possibly
    abbreviated) mode/class and length (as \code{\var{type}[1:\dots]}).}

\item{give.length}{logical; if \code{TRUE} (default), indicate
    length (as \code{[1:\dots]}).}

\item{...}{potential further arguments (required for Method/Generic
    reasons).}
}
\value{
\code{\link{invisible}}
}
\description{
To actually view the internal structure use \code{str(unclass(bit))}
}
\examples{
str(bit(120))
}
