% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZBTB7A_Demo}
\alias{ZBTB7A_Demo}
\title{ZBTB7A Demo data}
\format{
## 'ZBTB7A_Demo'
A data frame with 13511 rows and 4 columns
\describe{
  \item{TF}{transcription factor HGNC symbol.}
  \item{GOOD}{number of good cases of demo input ATAC-seq peakset.}
  \item{TOTAL}{number of total informative cases of demo input ATAC-seq peakset.}
  \item{Ratio}{ratio of good cases over all informative cases.}
}
}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE173416
}
\usage{
ZBTB7A_Demo
}
\description{
A transformed demo data set used for showing the workflow of BIMTR.
}
\details{
King AJ, Songdej D, Downes DJ, Beagrie RA, Liu S, Buckley M, Hua P, Suciu MC, Marieke Oudelaar A, Hanssen LLP, Jeziorska D,
Roberts N, Carpenter SJ, Francis H, Telenius J, Olijnik AA, Sharpe JA, Sloane-Stanley J, Eglinton J, Kassouf MT, Orkin SH,
Pennacchio LA, Davies JOJ, Hughes JR, Higgs DR, Babbs C.
Reactivation of a developmentally silenced embryonic globin gene. Nat Commun. 2021 Jul 21;12(1):4439.
doi: 10.1038/s41467-021-24402-3. PMID: 34290235; PMCID: PMC8295333.
}
\keyword{datasets}
