% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{BayesIMTR}
\alias{BayesIMTR}
\title{BayesIMTR}
\usage{
BayesIMTR(
  file,
  format = c("bed", "narrowPeak", "broadPeak", "bigNarrowPeak"),
  N = 2000,
  bin_width = 1000,
  version = c(1, 2),
  centerized = FALSE
)
}
\arguments{
\item{file}{file path to the user-input.}

\item{format}{format can be .bed, .narrowPeak, .broadPeak and .bigNarrowPeak.}

\item{N}{number of rounds for gibbs sampler, recommended for > 2000.}

\item{bin_width}{desired width of bin, should be in 100/500/1000.}
}
\value{
A list object contains results of Gibbs sampler.
}
\description{
Main interface to run BayesIMTR method, user need to give the input file path, input file format, number of rounds and bin width.
users can also change default parameters used in Gibbs sampler.
}
