% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit-package.R
\name{bit-package}
\alias{bit-package}
\title{bit: Classes and methods for fast memory-efficient boolean selections}
\description{
Provided are classes for boolean and skewed boolean vectors, fast boolean
methods, fast unique and non-unique integer sorting, fast set operations on
sorted and unsorted sets of integers, and foundations for ff (range indices,
compression, chunked processing).
}
\details{
For details view the \code{vignette("bit-usage")} and \code{vignette("bit-performance")}.
}
