% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{BIT}
\alias{BIT}
\title{BIT}
\usage{
BIT(
  file,
  format = c("bed", "narrowPeak", "broadPeak", "bigNarrowPeak", "csv"),
  N = 5000,
  bin_width = c(100, 500, 1000),
  option = c("ALL", "CREs", "PLS", "ELS")
)
}
\arguments{
\item{file}{file path to the user-input.}

\item{format}{format can be .bed, .narrowPeak, .broadPeak and .bigNarrowPeak.}

\item{N}{number of iterations for gibbs sampler, recommended for >= 5000.}

\item{bin_width}{desired width of bin, should be in 100/500/1000.}

\item{option}{option to filter peaks with candidate cis-regulatory elements from ENCODE.}
}
\value{
A list object contains results of Gibbs sampler.
}
\description{
Main interface to run BIT method, please set the input file path, input file format, number of iterations and bin width.
users can also change default parameters used in Gibbs sampler.
}
