% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{BIT_compare}
\alias{BIT_compare}
\title{BIT_compare}
\usage{
BIT_compare(
  file1,
  file2,
  output_path,
  show = TRUE,
  plot_scatter = TRUE,
  format = c(NULL, NULL),
  N = 5000,
  bin_width = 1000,
  burnin = NULL,
  genome = c("hg38", "mm10")
)
}
\arguments{
\item{file1}{file path to the user-input file 1.}

\item{file2}{file path to the user-input file 2.}

\item{output_path}{absoluate or relative directory to store the Gibbs sampler data.}

\item{format}{if specify as NULL, BIT will automatically read and judge the file type based on extension, default: NULL.}

\item{N}{number of iterations for gibbs sampler, recommended for >= 5000, default: 5000.}

\item{bin_width}{desired width of bin, should be in 100/500/1000, default: 1000.}

\item{option}{option to filter peaks with candidate cis-regulatory elements from ENCODE,
default use "ALL", can be "PLS" for all promoters, "ELS" for all enhancers.}
}
\description{
compare BIT identifid TRs for two user input epigenomic region sets.
}
