% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demo.R
\name{Demo}
\alias{Demo}
\title{Demo to show BayesIMTR, which can be used without loading the ChIP-seq data.}
\usage{
Demo(N, dat = c("CTCF", "ZBTB7A", "KDM1A"))
}
\arguments{
\item{N}{number of rounds for gibbs sampler, recommended for > 2000.}

\item{dat}{pre-processed Demo data, including CTCF, ZBTB7A and KDM1A. For details please check the manual for the data.}
}
\value{
A list object contains results of Gibbs sampler.
}
\description{
Demo to show BayesIMTR, which can be used without loading the ChIP-seq data.
}
