% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R
\name{is.na.bit}
\alias{is.na.bit}
\alias{is.na.bitwhich}
\title{Test for NA in bit and bitwhich}
\usage{
\method{is.na}{bit}(x)

\method{is.na}{bitwhich}(x)
}
\arguments{
\item{x}{a \code{\link[=bit]{bit()}} or  \code{\link[=bitwhich]{bitwhich()}} vector}
}
\value{
vector of same type with all elements \code{FALSE}
}
\description{
Test for NA in bit and bitwhich
}
\section{Functions}{
\itemize{
\item \code{is.na(bitwhich)}: method for \code{\link[=is.na]{is.na()}} from \code{\link[=bitwhich]{bitwhich()}}

}}
\examples{
is.na(bit(6))
is.na(bitwhich(6))
}
\seealso{
\code{\link[=is.na]{is.na()}}
}
