% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit.R
\name{ri}
\alias{ri}
\alias{print.ri}
\title{Range index}
\usage{
ri(from, to = NULL, maxindex = NA)

\method{print}{ri}(x, ...)
}
\arguments{
\item{from}{first position}

\item{to}{last posistion}

\item{maxindex}{the maximal length of the object-to-be-subscripted (if
known)}

\item{x}{an object of class 'ri'}

\item{...}{further arguments}
}
\value{
A two element integer vector with class 'ri'
}
\description{
A range index can be used to extract or replace a continuous ascending part
of the data
}
\examples{

 bit(12)[ri(1,6)]

}
\seealso{
\code{\link[ff:as.hi]{ff::as.hi()}}
}
\author{
Jens Oehlschlägel
}
\keyword{classes}
\keyword{logic}
