% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demo.R
\name{Show_Results}
\alias{Show_Results}
\title{To show the ranking table by inspecting the results of Gibbs sampler.}
\usage{
Show_Results(dat, burnin)
}
\arguments{
\item{dat}{results from Main_sampling.}

\item{burnin}{number of samples used for burn-in.}
}
\value{
a list object contains two data.frame. one by ranking the theta_ij, the other one by ranking theta_i.
}
\description{
To show the ranking table by inspecting the results of Gibbs sampler.
}
\examples{
CTCF_Demo <- Demo(1000, "CTCF")
CTCF_Results <- Show_Results(CTCF_Demo, burnin = 500)
head(CTCF_Results[["Theta_ij"]],10)
head(CTCF_Results[["Theta_i"]],10)
}
