% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrutil.R
\name{unattr}
\alias{unattr}
\title{Attribute removal}
\usage{
unattr(x)
}
\arguments{
\item{x}{any R object}
}
\value{
a similar object with attributes removed
}
\description{
Returns object with attributes removed
}
\details{
attribute removal copies the object as usual
}
\examples{

  bit(2)[]
  unattr(bit(2)[])

}
\seealso{
\code{\link{attributes}}, \code{\link{setattributes}},
\code{\link{unclass}}
}
\author{
Jens Oehlschlägel
}
\keyword{attribute}
